+++ src/main/resources/templates/Admin_Dashboard/Members_Search.html
@@ -97,11 +97,11 @@
                         <h6 style="font-size: 15px; color:gray; text-align:center">총 관리자 권한이 있는 대상만 회원 권한부여/조회 할 수 있습니다.
                         </h6>
                         <br>
-                        <form class="form-inline">
+                        <form class="form-inline" method="get" action="/Members_Search">
                             <div class="input-group">
-                                <input class="form-control" type="search"
+                                <input class="form-control" type="search" name="searchValue"
                                     placeholder="회원의 이메일 또는 닉네임으로 검색하여 추가할 수 있습니다." aria-label="Search">
-                                <button class="btn btn-outline-primary" type="button">
+                                <button class="btn btn-outline-primary" type="submit">
                                     <i class="fas fa-search"></i>
                                 </button>
                             </div>
